Day 3/Demo 3/requirements.txt
azure-functions==1.21.3
psycopg2-binary
debugpy
